Command関連をクラス化完了
