+++ Adapter/component/Resource.cpp
@@ -4,15 +4,15 @@ using namespace LWP::System;
 using namespace LWP;
 
 Resource::Texture* Resource::LoadTexture(const std::string& filePath) {
-	return engine->resourceManager_->LoadTexture(engine->commandManager_.get(), filePath);
+	return engine->resourceManager_->LoadTexture(engine->directXCommon_->GetCommandManager(), filePath);
 }
 Resource::Texture* Resource::LoadTextureLongPath(const std::string& filePath) {
-	return engine->resourceManager_->LoadTextureLongPath(engine->commandManager_.get(), filePath);
+	return engine->resourceManager_->LoadTextureLongPath(engine->directXCommon_->GetCommandManager(), filePath);
 }
 
 
 Primitive::Mesh* Resource::LoadModel(const std::string& filePath) {
-	Primitive::Mesh* instance = engine->primitiveManager_->CreateInstance<Primitive::Mesh>(engine->commandManager_.get());
+	Primitive::Mesh* instance = engine->primitiveManager_->CreateInstance<Primitive::Mesh>(engine->directXCommon_->GetCommandManager());
 	instance->LoadFile(filePath);
 	return instance;
 }
