+++ DirectXGame.vcxproj
@@ -105,9 +105,9 @@ copy "$(WindowsSdkDir)bin\$(TargetPlatformVersion)\x64\dxil.dll" "$(TargetDir)dx
     <ClCompile Include="Class\Mob\Derived\Enemy.cpp" />
     <ClCompile Include="Class\Mob\IMob.cpp" />
     <ClCompile Include="Class\Mob\MobManager.cpp" />
-    <ClCompile Include="Engine\base\CommandManager.cpp" />
     <ClCompile Include="Engine\base\DirectXCommon.cpp" />
-    <ClCompile Include="Engine\base\directX\Command.cpp" />
+    <ClCompile Include="Engine\base\directX\command\Command.cpp" />
+    <ClCompile Include="Engine\base\directX\command\CommandManager.cpp" />
     <ClCompile Include="Engine\base\directX\descriptorHeap\DSV.cpp" />
     <ClCompile Include="Engine\base\directX\descriptorHeap\IDescriptorHeap.cpp" />
     <ClCompile Include="Engine\base\directX\descriptorHeap\RTV.cpp" />
@@ -173,9 +173,10 @@ copy "$(WindowsSdkDir)bin\$(TargetPlatformVersion)\x64\dxil.dll" "$(TargetDir)dx
     <ClInclude Include="Class\Mob\Derived\Enemy.h" />
     <ClInclude Include="Class\Mob\IMob.h" />
     <ClInclude Include="Class\Mob\MobManager.h" />
-    <ClInclude Include="Engine\base\CommandManager.h" />
     <ClInclude Include="Engine\base\DirectXCommon.h" />
-    <ClInclude Include="Engine\base\directX\Command.h" />
+    <ClInclude Include="Engine\base\directX\command\Command.h" />
+    <ClInclude Include="Engine\base\directX\command\CommandManager.h" />
+    <ClInclude Include="Engine\base\directX\command\ResourceStruct.h" />
     <ClInclude Include="Engine\base\directX\descriptorHeap\DSV.h" />
     <ClInclude Include="Engine\base\directX\descriptorHeap\IDescriptorHeap.h" />
     <ClInclude Include="Engine\base\directX\descriptorHeap\RTV.h" />
