+++ DirectXGame.vcxproj.filters
@@ -58,9 +58,6 @@
     <ClCompile Include="Externals\imgui\imgui_widgets.cpp">
       <Filter>Externals\imgui</Filter>
     </ClCompile>
-    <ClCompile Include="Engine\base\CommandManager.cpp">
-      <Filter>Engine\base</Filter>
-    </ClCompile>
     <ClCompile Include="Engine\base\WinApp.cpp">
       <Filter>Engine\base</Filter>
     </ClCompile>
@@ -184,8 +181,11 @@
     <ClCompile Include="Engine\base\directX\descriptorHeap\DSV.cpp">
       <Filter>Engine\base\directX\descriptorHeap</Filter>
     </ClCompile>
-    <ClCompile Include="Engine\base\directX\Command.cpp">
-      <Filter>Engine\base\directX</Filter>
+    <ClCompile Include="Engine\base\directX\command\Command.cpp">
+      <Filter>Engine\base\directX\command</Filter>
+    </ClCompile>
+    <ClCompile Include="Engine\base\directX\command\CommandManager.cpp">
+      <Filter>Engine\base\directX\command</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -224,9 +224,6 @@
     </ClInclude>
     <ClInclude Include="Externals\imgui\imgui_internal.h">
       <Filter>Externals\imgui</Filter>
-    </ClInclude>
-    <ClInclude Include="Engine\base\CommandManager.h">
-      <Filter>Engine\base</Filter>
     </ClInclude>
     <ClInclude Include="Engine\base\WinApp.h">
       <Filter>Engine\base</Filter>
@@ -387,8 +384,14 @@
     <ClInclude Include="Engine\base\directX\descriptorHeap\DSV.h">
       <Filter>Engine\base\directX\descriptorHeap</Filter>
     </ClInclude>
-    <ClInclude Include="Engine\base\directX\Command.h">
-      <Filter>Engine\base\directX</Filter>
+    <ClInclude Include="Engine\base\directX\command\Command.h">
+      <Filter>Engine\base\directX\command</Filter>
+    </ClInclude>
+    <ClInclude Include="Engine\base\directX\command\CommandManager.h">
+      <Filter>Engine\base\directX\command</Filter>
+    </ClInclude>
+    <ClInclude Include="Engine\base\directX\command\ResourceStruct.h">
+      <Filter>Engine\base\directX\command</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -499,6 +502,9 @@
     </Filter>
     <Filter Include="Engine\base\directX\descriptorHeap">
       <UniqueIdentifier>{89f131fc-3cd1-4d6d-b6a7-a88f69bf7b16}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Engine\base\directX\command">
+      <UniqueIdentifier>{60100a09-802c-4da4-904f-ccf44e31b290}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
 </Project>
