--- Engine/base/directX/Command.cpp
@@ -1,16 +1,0 @@
-#include "Command.h"
-
-using namespace LWP::Base;
-
-
-void Command::Initialize() {
-
-}
-
-void Command::PreDraw() {
-
-}
-
-void Command::PostDraw() {
-
-}
