+++ imgui.ini
@@ -59,7 +59,7 @@ Size=212,145
 Collapsed=0
 
 [Window][Primitive]
-Pos=34,19
+Pos=36,19
 Size=294,457
 Collapsed=0
